ReportService unittest edited, repor method added

+++ GuideBook.Dto/ExcelReportViewModel.cs
@@ -5,7 +5,7 @@ public class ExcelReportViewModel
     public ExcelReportViewModel(string location) => Location = location;
     public ExcelReportViewModel() { }
     
-    public int PersonCount { get; set; }
+    public long PersonCount { get; set; }
     public string Location { get; set; }
-    public int PhoneCount { get; set; }
+    public long PhoneCount { get; set; }
 }
